Move GitHub classes into their own module
